app/requirements.txt
asyncio==3.4.3
psycopg2==2.9.5
python-dotenv==0.21.1
websockets==10.4
